+++ src/main/java/com/coolerpromc/bettercampfirepot/menu/CookingPotMenu.java
@@ -19,7 +19,7 @@ public class CookingPotMenu extends AbstractContainerMenu {
     public final BetterCampfireBlockEntity blockEntity;
 
     public CookingPotMenu(int containerId, Inventory playerInventory, BlockPos blockPos) {
-        this(containerId, playerInventory, (BetterCampfireBlockEntity) playerInventory.player.level().getBlockEntity(blockPos), new SimpleContainerData(4));
+        this(containerId, playerInventory, (BetterCampfireBlockEntity) playerInventory.player.level().getBlockEntity(blockPos), new SimpleContainerData(5));
     }
 
     public CookingPotMenu(int containerId, Inventory playerInventory, BetterCampfireBlockEntity blockEntity, ContainerData containerData){