+++ src/main/java/com/coolerpromc/bettercampfirepot/menu/CookingPotScreen.java
@@ -3,15 +3,23 @@ package com.coolerpromc.bettercampfirepot.menu;
 import com.coolerpromc.bettercampfirepot.BetterCampfirePot;
 import com.coolerpromc.bettercampfirepot.item.BetterCampfirePotItem;
 import com.coolerpromc.bettercampfirepot.menu.widget.BetterCookButton;
+import com.coolerpromc.bettercampfirepot.menu.widget.ChangeCapabilityButton;
+import com.coolerpromc.bettercampfirepot.menu.widget.ToggleConfigButton;
+import com.coolerpromc.bettercampfirepot.network.CapabilityChangeSyncC2SPacket;
 import com.coolerpromc.bettercampfirepot.network.ToggleCookingPotLidPacket;
+import com.coolerpromc.bettercampfirepot.util.Side;
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.FastColor;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.Slot;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static com.cobblemon.mod.common.util.MiscUtilsKt.cobblemonResource;
 import static com.coolerpromc.bettercampfirepot.block.entity.BetterCampfireBlockEntity.IS_LID_OPEN_INDEX;
@@ -21,6 +29,8 @@ public class CookingPotScreen extends AbstractContainerScreen<CookingPotMenu> {
     private final ResourceLocation COOK_PROGRESS_SPRITE = cobblemonResource("textures/gui/campfirepot/cook_progress.png");
 
     private BetterCookButton cookButton;
+    private ToggleConfigButton toggleConfigButton;
+    private final List<ChangeCapabilityButton> capabilityButtons = new ArrayList<>();
 
     public CookingPotScreen(CookingPotMenu menu, Inventory playerInventory, Component title) {
         super(menu, playerInventory, title);
@@ -41,6 +51,39 @@ public class CookingPotScreen extends AbstractContainerScreen<CookingPotMenu> {
             });
             addRenderableWidget(cookButton);
         }
+
+        toggleConfigButton = new ToggleConfigButton(leftPos + imageWidth + 4, topPos + 25, 20, 20, button -> {
+            if (button.isToggled()){
+                openConfigScreen();
+            }
+            else{
+                closeConfigScreen();
+            }
+        });
+        addRenderableWidget(toggleConfigButton);
+
+        int x = leftPos + imageWidth + 1;
+        int y = topPos + 49;
+
+        capabilityButtons.add(new ChangeCapabilityButton(x + 9, y + 4, 8, 8, this::onPress, Side.TOP, this.menu.blockEntity.getCapabilityBySide(Side.TOP)));
+        capabilityButtons.add(new ChangeCapabilityButton(x + 1, y + 12, 8, 8, this::onPress, Side.LEFT, this.menu.blockEntity.getCapabilityBySide(Side.LEFT)));
+        capabilityButtons.add(new ChangeCapabilityButton(x + 9, y + 12, 8, 8, this::onPress, Side.FRONT, this.menu.blockEntity.getCapabilityBySide(Side.FRONT)));
+        capabilityButtons.add(new ChangeCapabilityButton(x + 17, y + 12, 8, 8, this::onPress, Side.RIGHT, this.menu.blockEntity.getCapabilityBySide(Side.RIGHT)));
+        capabilityButtons.add(new ChangeCapabilityButton(x + 1, y + 20, 8, 8, this::onPress, Side.BACK, this.menu.blockEntity.getCapabilityBySide(Side.BACK)));
+        capabilityButtons.add(new ChangeCapabilityButton(x + 9, y + 20, 8, 8, this::onPress, Side.BOTTOM, this.menu.blockEntity.getCapabilityBySide(Side.BOTTOM)));
+    }
+
+    private void onPress(ChangeCapabilityButton button){
+        button.setSlot(button.getSlot().next());
+        ClientPlayNetworking.send(new CapabilityChangeSyncC2SPacket(this.menu.blockEntity.getBlockPos(), button.getSide(), button.getSlot()));
+    }
+
+    private void openConfigScreen(){
+        this.capabilityButtons.forEach(this::addRenderableWidget);
+    }
+
+    private void closeConfigScreen(){
+        this.capabilityButtons.forEach(this::removeWidget);
     }
 
     @Override
@@ -48,6 +91,10 @@ public class CookingPotScreen extends AbstractContainerScreen<CookingPotMenu> {
         guiGraphics.blit(BACKGROUND, leftPos, topPos, 0, 0, imageWidth, imageHeight, 176, 166);
         int cookProgress = (int) Math.ceil(menu.getBurnProgress() * 22);
         guiGraphics.blit(COOK_PROGRESS_SPRITE, leftPos + 96, topPos + 39, 0, 0, cookProgress, 12, 22, 12);
+        blitWithBorder(guiGraphics, BetterCampfirePot.id("textures/gui/sprites/gui.png"), leftPos + imageWidth - 1, topPos + 20, 0, 0, 30, 30, 24, 24, 4);
+        if (toggleConfigButton.isToggled()){
+            blitWithBorder(guiGraphics, BetterCampfirePot.id("textures/gui/sprites/gui.png"), leftPos + imageWidth - 1, topPos + 50, 24, 0, 30, 30, 24, 24, 4);
+        }
     }
 
     @Override
@@ -78,6 +125,7 @@ public class CookingPotScreen extends AbstractContainerScreen<CookingPotMenu> {
             }
         }
 
+        renderSlotOutline(guiGraphics);
 
         renderTooltip(guiGraphics, mouseX, mouseY);
     }
@@ -90,5 +138,88 @@ public class CookingPotScreen extends AbstractContainerScreen<CookingPotMenu> {
             guiGraphics.fill(x,y,x+16,y+16,822018048);
         }
         super.renderSlot(guiGraphics, slot);
+    }
+
+    public boolean isConfigOpened(){
+        return toggleConfigButton.isToggled();
+    }
+
+    private void renderSlotOutline(GuiGraphics guiGraphics){
+        for (ChangeCapabilityButton btn : capabilityButtons){
+            if (btn.isHovered()){
+                int baseX = leftPos;
+                int baseY = topPos;
+                int maxX = baseX;
+                int maxY = baseY;
+
+                switch (btn.getSlot()){
+                    case INPUT:
+                        baseX += 32;
+                        baseY += 17;
+                        maxX += 54 + 32;
+                        maxY += 54 + 17;
+                        break;
+                    case SEASONING:
+                        baseX += 109;
+                        baseY += 17;
+                        maxX += 54 + 109;
+                        maxY += 18 + 17;
+                        break;
+                    case OUTPUT:
+                        baseX += 123;
+                        baseY += 50;
+                        maxX += 26 + 123;
+                        maxY += 26 + 50;
+                        break;
+                }
+
+                guiGraphics.fill(baseX, baseY, maxX, maxY, FastColor.ARGB32.color(0x66, btn.getSlot().color()));
+            }
+        }
+    }
+
+    public int getGuiLeft(){
+        return leftPos;
+    }
+
+    public int getGuiTop(){
+        return topPos;
+    }
+
+    public int getXSize(){
+        return imageWidth;
+    }
+
+    private void blitWithBorder(GuiGraphics guiGraphics, ResourceLocation texture, int x, int y, int u, int v, int width, int height, int textureWidth, int textureHeight, int borderSize) {
+        this.blitWithBorder(guiGraphics, texture, x, y, u, v, width, height, textureWidth, textureHeight, borderSize, borderSize, borderSize, borderSize);
+    }
+
+    private void blitWithBorder(GuiGraphics guiGraphics, ResourceLocation texture, int x, int y, int u, int v, int width, int height, int textureWidth, int textureHeight, int topBorder, int bottomBorder, int leftBorder, int rightBorder) {
+        int fillerWidth = textureWidth - leftBorder - rightBorder;
+        int fillerHeight = textureHeight - topBorder - bottomBorder;
+        int canvasWidth = width - leftBorder - rightBorder;
+        int canvasHeight = height - topBorder - bottomBorder;
+        int xPasses = canvasWidth / fillerWidth;
+        int remainderWidth = canvasWidth % fillerWidth;
+        int yPasses = canvasHeight / fillerHeight;
+        int remainderHeight = canvasHeight % fillerHeight;
+        guiGraphics.blit(texture, x, y, u, v, leftBorder, topBorder);
+        guiGraphics.blit(texture, x + leftBorder + canvasWidth, y, u + leftBorder + fillerWidth, v, rightBorder, topBorder);
+        guiGraphics.blit(texture, x, y + topBorder + canvasHeight, u, v + topBorder + fillerHeight, leftBorder, bottomBorder);
+        guiGraphics.blit(texture, x + leftBorder + canvasWidth, y + topBorder + canvasHeight, u + leftBorder + fillerWidth, v + topBorder + fillerHeight, rightBorder, bottomBorder);
+
+        for(int i = 0; i < xPasses + (remainderWidth > 0 ? 1 : 0); ++i) {
+            guiGraphics.blit(texture, x + leftBorder + i * fillerWidth, y, u + leftBorder, v, i == xPasses ? remainderWidth : fillerWidth, topBorder);
+            guiGraphics.blit(texture, x + leftBorder + i * fillerWidth, y + topBorder + canvasHeight, u + leftBorder, v + topBorder + fillerHeight, i == xPasses ? remainderWidth : fillerWidth, bottomBorder);
+
+            for(int j = 0; j < yPasses + (remainderHeight > 0 ? 1 : 0); ++j) {
+                guiGraphics.blit(texture, x + leftBorder + i * fillerWidth, y + topBorder + j * fillerHeight, u + leftBorder, v + topBorder, i == xPasses ? remainderWidth : fillerWidth, j == yPasses ? remainderHeight : fillerHeight);
+            }
+        }
+
+        for(int j = 0; j < yPasses + (remainderHeight > 0 ? 1 : 0); ++j) {
+            guiGraphics.blit(texture, x, y + topBorder + j * fillerHeight, u, v + topBorder, leftBorder, j == yPasses ? remainderHeight : fillerHeight);
+            guiGraphics.blit(texture, x + leftBorder + canvasWidth, y + topBorder + j * fillerHeight, u + leftBorder + fillerWidth, v + topBorder, rightBorder, j == yPasses ? remainderHeight : fillerHeight);
+        }
     }
 }